Add option to preserve variables in output
Optionally preserve variables in the output, with the resolved values
inserted before the variable reference. This makes it possible to use
the CSS custom properties in supporting browsers, and have the resolved
values act as fallbacks in non-supporting browsers.

This changes the signature of the plugin. Variables passed in via JS
must now be provided as an object assigned to the `map` property of the
plugin options. By default, `preserve` is `false`.

options.map: Object
options.preserve: Boolean

Fix #28